[HttpKernel] Add a ContainerControllerResolver (psr-11)
