+++ src/Symfony/Bundle/FrameworkBundle/Controller/ControllerResolver.php
@@ -12,18 +12,17 @@
 namespace Symfony\Bundle\FrameworkBundle\Controller;
 
 use Psr\Log\LoggerInterface;
-use Symfony\Component\HttpKernel\Controller\ControllerResolver as BaseControllerResolver;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
+use Symfony\Component\HttpKernel\Controller\ContainerControllerResolver;
 
 /**
  * ControllerResolver.
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class ControllerResolver extends BaseControllerResolver
+class ControllerResolver extends ContainerControllerResolver
 {
-    protected $container;
     protected $parser;
 
     /**
@@ -35,39 +34,19 @@ class ControllerResolver extends BaseControllerResolver
      */
     public function __construct(ContainerInterface $container, ControllerNameParser $parser, LoggerInterface $logger = null)
     {
-        $this->container = $container;
         $this->parser = $parser;
 
-        parent::__construct($logger);
+        parent::__construct($container, $logger);
     }
 
     /**
-     * Returns a callable for the given controller.
-     *
-     * @param string $controller A Controller string
-     *
-     * @return mixed A PHP callable
-     *
-     * @throws \LogicException           When the name could not be parsed
-     * @throws \InvalidArgumentException When the controller class does not exist
+     * {@inheritdoc}
      */
     protected function createController($controller)
     {
-        if (false === strpos($controller, '::')) {
-            $count = substr_count($controller, ':');
-            if (2 == $count) {
-                // controller in the a:b:c notation then
-                $controller = $this->parser->parse($controller);
-            } elseif (1 == $count) {
-                // controller in the service:method notation
-                list($service, $method) = explode(':', $controller, 2);
-
-                return array($this->container->get($service), $method);
-            } elseif ($this->container->has($controller) && method_exists($service = $this->container->get($controller), '__invoke')) {
-                return $service;
-            } else {
-                throw new \LogicException(sprintf('Unable to parse the controller name "%s".', $controller));
-            }
+        if (false === strpos($controller, '::') && 2 === substr_count($controller, ':')) {
+            // controller in the a:b:c notation then
+            $controller = $this->parser->parse($controller);
         }
 
         return parent::createController($controller);
@@ -78,10 +57,6 @@ class ControllerResolver extends BaseControllerResolver
      */
     protected function instantiateController($class)
     {
-        if ($this->container->has($class)) {
-            return $this->container->get($class);
-        }
-
         $controller = parent::instantiateController($class);
 
         if ($controller instanceof ContainerAwareInterface) {
